+++ python/beeai_framework/adapters/a2a/agents/agent.py
@@ -1,16 +1,5 @@
 # Copyright 2025 © BeeAI a Series of LF Projects, LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
+# SPDX-License-Identifier: Apache-2.0
 
 from collections.abc import AsyncGenerator
 from uuid import uuid4
@@ -97,7 +86,7 @@ class A2AAgent(BaseAgent[A2AAgentRunOutput]):
                 )
 
                 streaming_request: a2a_types.SendStreamingMessageRequest = a2a_types.SendStreamingMessageRequest(
-                    params=a2a_types.MessageSendParams(message=self._convert_to_a2a_message(input))
+                    id=str(uuid4().hex), params=a2a_types.MessageSendParams(message=self._convert_to_a2a_message(input))
                 )
 
                 last_event_with_data = None