+++ python/beeai_framework/serve/server.py
@@ -1,16 +1,5 @@
 # Copyright 2025 © BeeAI a Series of LF Projects, LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
+# SPDX-License-Identifier: Apache-2.0
 
 from abc import ABC, abstractmethod
 from collections.abc import Callable, Sequence
@@ -18,6 +7,8 @@ from typing import ClassVar, Generic, Self
 
 from pydantic import BaseModel
 from typing_extensions import TypeVar
+
+from beeai_framework.serve.errors import FactoryAlreadyRegisteredError
 
 TInput = TypeVar("TInput", bound=object, default=object, contravariant=True)
 TInternal = TypeVar("TInternal", bound=object, default=object)
@@ -43,7 +34,7 @@ class Server(Generic[TInput, TInternal, TConfig], ABC):
         if ref not in cls._factories or override:
             cls._factories[ref] = factory
         elif cls._factories[ref] is not factory:
-            raise ValueError(f"Factory for {ref} is already registered.")
+            raise FactoryAlreadyRegisteredError(f"Factory for {ref} is already registered.")
 
     def register(self, input: TInput) -> Self:
         # check if the type has a factory registered