+++ python/cz_commitizen/monorepo_commits.py
@@ -1,16 +1,5 @@
 # Copyright 2025 © BeeAI a Series of LF Projects, LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
+# SPDX-License-Identifier: Apache-2.0
 
 from collections.abc import Iterable
 from typing import Any
@@ -20,7 +9,7 @@ from commitizen.cz.conventional_commits import ConventionalCommitsCz  # type: ig
 
 __all__ = ["MonorepoCommitsCz"]
 
-from commitizen.defaults import Questions
+from commitizen.question import CzQuestion
 
 
 class MonorepoCommitsCz(ConventionalCommitsCz):
@@ -43,10 +32,10 @@ class MonorepoCommitsCz(ConventionalCommitsCz):
         parent_hook = super().changelog_message_builder_hook
         return parent_hook(parsed_message, commit) if parent_hook else parsed_message
 
-    def questions(self) -> Questions:
+    def questions(self) -> list[CzQuestion]:
         questions = super().questions()
         for index, question in enumerate(questions):
-            if question["name"] == "prefix":
+            if question["type"] == "list" and question["name"] == "prefix":
                 question["choices"].append({"value": "chore", "name": "chore: other uncategorized changes"})
 
             if question["name"] == "scope":